Initial
This is working code that involves the client sending 10 files to the server by sending the next 1-3 lines of a random file until all lines of all files have been sent. Then, the server combines all the input into one file with the files' content separated from each other and in order.